03-natural-language/question-answering/requirements.txt
azure-ai-language-questionanswering==1.1.0
azure-core==1.28.0
python-dotenv==1.0.0
requests==2.31.0